Moved and renamed HTML/Unicode to Latex converters
