+++ src/main/java/net/sf/jabref/importer/fetcher/ACMPortalFetcher.java
@@ -19,10 +19,10 @@ package net.sf.jabref.importer.fetcher;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.gui.FetcherPreviewDialog;
-import net.sf.jabref.importer.HTMLConverter;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.importer.fileformat.BibtexParser;
+import net.sf.jabref.logic.formatter.bibtexfields.HTMLToLatexFormatter;
 import net.sf.jabref.logic.formatter.bibtexfields.UnitFormatter;
 import net.sf.jabref.logic.formatter.casechanger.CaseKeeper;
 import net.sf.jabref.logic.l10n.Localization;
@@ -53,7 +53,7 @@ public class ACMPortalFetcher implements PreviewEntryFetcher {
 
     private static final Log LOGGER = LogFactory.getLog(ACMPortalFetcher.class);
 
-    private final HTMLConverter htmlConverter = new HTMLConverter();
+    private final HTMLToLatexFormatter htmlConverter = new HTMLToLatexFormatter();
     private final CaseKeeper caseKeeper = new CaseKeeper();
     private final UnitFormatter unitFormatter = new UnitFormatter();
     private String terms;
