+++ src/main/java/net/sf/jabref/logic/formatter/BibtexFieldFormatters.java
@@ -1,7 +1,6 @@
 package net.sf.jabref.logic.formatter;
 
 import net.sf.jabref.logic.formatter.bibtexfields.*;
-
 import java.util.Arrays;
 import java.util.List;
 
@@ -15,6 +14,10 @@ public class BibtexFieldFormatters {
     public static final UnitFormatter UNIT_FORMATTER = new UnitFormatter();
     public static final TrimFormatter TRIM_FORMATTER = new TrimFormatter();
     public static final RemoveBracesFormatter REMOVE_BRACES_FORMATTER = new RemoveBracesFormatter();
+    public static final HTMLToLatexFormatter HTML_TO_LATEX = new HTMLToLatexFormatter();
+    public static final UnicodeToLatexFormatter UNICODE_TO_LATEX = new UnicodeToLatexFormatter();
 
-    public static final List<Formatter> ALL = Arrays.asList(PAGE_NUMBERS, SUPERSCRIPTS, DATE, AUTHORS_FORMATTER, LATEX_FORMATTER, MONTH_FORMATTER, UNIT_FORMATTER, TRIM_FORMATTER, REMOVE_BRACES_FORMATTER);
+    public static final List<Formatter> ALL = Arrays.asList(PAGE_NUMBERS, SUPERSCRIPTS, DATE, AUTHORS_FORMATTER,
+            LATEX_FORMATTER, MONTH_FORMATTER, UNIT_FORMATTER, TRIM_FORMATTER, REMOVE_BRACES_FORMATTER, HTML_TO_LATEX,
+            UNICODE_TO_LATEX);
 }
