+++ src/main/java/net/sf/jabref/logic/formatter/bibtexfields/UnicodeToLatexFormatter.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package net.sf.jabref.importer;
+package net.sf.jabref.logic.formatter.bibtexfields;
 
 import java.util.Objects;
 import java.util.Set;
@@ -24,12 +24,12 @@ import net.sf.jabref.exporter.layout.LayoutFormatter;
 import net.sf.jabref.logic.formatter.Formatter;
 import net.sf.jabref.logic.util.strings.HTMLUnicodeConversionMaps;
 
-public class UnicodeConverter implements LayoutFormatter, Formatter {
+public class UnicodeToLatexFormatter implements LayoutFormatter, Formatter {
 
-    private static final Log LOGGER = LogFactory.getLog(UnicodeConverter.class);
+    private static final Log LOGGER = LogFactory.getLog(UnicodeToLatexFormatter.class);
 
 
-    public UnicodeConverter() {
+    public UnicodeToLatexFormatter() {
         super();
     }
