+++ src/test/java/net/sf/jabref/logic/formatter/FormatterTest.java
@@ -2,7 +2,6 @@ package net.sf.jabref.logic.formatter;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.importer.HTMLConverter;
 import net.sf.jabref.logic.formatter.bibtexfields.*;
 import net.sf.jabref.logic.formatter.casechanger.*;
 import net.sf.jabref.logic.formatter.minifier.AuthorsMinifier;
@@ -82,7 +81,7 @@ public class FormatterTest {
                 new Object[]{new PageNumbersFormatter()},
                 new Object[]{new LowerCaseChanger()},
                 new Object[]{new TrimFormatter()},
-                new Object[]{new HTMLConverter()},
+                new Object[]{new HTMLToLatexFormatter()},
                 new Object[]{new SuperscriptFormatter()},
                 new Object[]{new UnitFormatter()},
                 new Object[]{new RemoveBracesFormatter()}
